requirements.txt
# Core
nicegui>=1.4.20,<2.0
sqlalchemy>=2.0.30,<3.0
pydantic>=2.7.0,<3.0
pydantic-settings>=2.2.0,<3.0
python-dotenv>=1.0.1
pyyaml>=6.0.1

# Exchange & Data
httpx>=0.27.0,<1.0
pandas>=2.2.0
numpy>=1.26.0,<3.0

# Async
aiohttp>=3.9.0
aiosqlite>=0.20.0

# Telegram (we use raw HTTP API; this is here for future use)
python-telegram-bot>=21.0

# Charts (optional — used by NiceGUI for plotly elements)
plotly>=5.20.0

# Logging
loguru>=0.7.2

# Scheduling
apscheduler>=3.10.4
